+++ sdk/merchantSession/log/log.js
@@ -3,12 +3,11 @@ const check = require('starkcore').check;
 const Resource = require('../../utils/resource.js').Resource
 
 class Log extends Resource {
-
     /**
      * Check out our API Documentation at https://starkbank.com/docs/api#merchant-session
      */
 
-    constructor({ created, type, errors, session, id }) {
+    constructor({created, type, errors, session, id}) {
         super(id);
         this.created = check.datetime(created);
         this.type = type;
@@ -24,21 +23,23 @@ exports.get = async function (id, {user} = {}) {
     return rest.getId(resource, id, user);
 };
 
-exports.query = async function ({ limit, after, before, user } = {}) {
+exports.query = async function ({limit, after, before, types, sessionIds, user} = {}) {
     let query = {
         limit: limit,
-        after: after,
-        before: before
+        after: check.date(after),
+        before: check.date(before),
+        types: types,
+        sessionIds: sessionIds,
     };
     return rest.getList(resource, query, user);
 };
 
-exports.page = async function ({ cursor, limit, after, before, types, sessionIds, user } = {}) {
+exports.page = async function ({cursor, limit, after, before, types, sessionIds, user} = {}) {
     let query = {
         cursor: cursor,
         limit: limit,
-        after: after,
-        before: before,
+        after: check.date(after),
+        before: check.date(before),
         types: types,
         sessionIds: sessionIds,
     };